Validation
added url validation.
added admin token validation for db access.
seperated routes.py from main.py
